+++ shared/opts-parser.ts
@@ -275,4 +275,3 @@ export namespace OptsParser {
       }).join('\n'));
   }
 }
-